requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
coloredlogs==15.0.1
cryptography==43.0.3
deepface==0.0.93
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-cloud-vision==3.9.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h5py==3.12.1
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka==1.3.5
kafka-python==2.0.2
keras==3.7.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
narwhals==1.21.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
pydeck==0.9.1
Pygments==2.19.1
PyMatting==1.1.13
PyMuPDF==1.24.14
pymupdf4llm==0.0.17
pyparsing==3.2.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
rembg==2.0.61
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-image==0.25.0
scipy==1.15.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tifffile==2024.12.12
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
